backend/monitor/monitor.requirements.txt
SQLAlchemy==1.3.20
psycopg2-binary==2.8.6
python-crontab==2.5.1
requests==2.22.0
configparser==3.5.3
